pyproject.toml
[project]
name = "agent-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "eth-account",
  "langchain",
  "langchain-google-genai",
  "langchain-community",
  "flask",
  "flask-cors",
  "python-dotenv"
]
